Refactor:
1)Make cubemap gl handle
2)Move enums to Defs.h

+++ engine/renderers/PBR.cpp
@@ -163,11 +163,11 @@ namespace rndr
 
 		//till we get sampler objects in
 		cubemap_bind(self->diffuse_irradiance_map, TEXTURE_UNIT::UNIT_0);
-		uniform1i_set(self->prog, "diffuse_irradiance_map", TEXTURE_UNIT::UNIT_0);
+		uniform1i_set(self->prog, "diffuse_irradiance_map", (int)TEXTURE_UNIT::UNIT_0);
 		cubemap_bind(self->specular_prefiltered_map, TEXTURE_UNIT::UNIT_1);
-		uniform1i_set(self->prog, "specular_prefiltered_map", TEXTURE_UNIT::UNIT_1);
+		uniform1i_set(self->prog, "specular_prefiltered_map", (int)TEXTURE_UNIT::UNIT_1);
 		texture2d_bind(self->specular_BRDF_LUT, TEXTURE_UNIT::UNIT_2);
-		uniform1i_set(self->prog, "specular_BRDF_LUT", TEXTURE_UNIT::UNIT_2);
+		uniform1i_set(self->prog, "specular_BRDF_LUT", (int)TEXTURE_UNIT::UNIT_2);
 
 		for (auto object : self->meshes)
 		{
